+++ docs/zh/.vitepress/config.js
@@ -44,7 +44,7 @@ module.exports = {
                     text: '进阶',
                     items: [
                         {text: '基于配置', link: '/guide/advance/options'},
-                        {text: '配置与组件的对应关系', link: '/guide/start/struct'},
+                        {text: '配置与组件的对应关系', link: '/guide/advance/struct'},
                         {text: '字段配置', link: '/guide/advance/component'},
                         {text: '字段类型', link: '/guide/advance/column-type'},
                         {text: '数据字典', link: '/guide/advance/dict'},