+++ packages/fast-crud/src/types/list/assist.ts
@@ -1,11 +1,11 @@
 import { uiContext } from "../../ui";
 
-function antdvColspan(ui, labelSpan) {
+function antdvColspan(ui: any, labelSpan: any) {
   return ui.type !== "antdv" ? {} : { labelCol: { span: labelSpan }, wrapperCol: { span: 23 - labelSpan } };
 }
 
 /**
- * 辅助type
+ * 辅助type 即将废弃，antdv已经有更好的方式做跨列配置
  */
 export default function () {
   const ui = uiContext.get();