+++ packages/fast-crud/src/use/use-drag.ts
@@ -25,7 +25,7 @@ async function dragModalForAntdv(opts: DragModalOptions) {
     //$(dragDom).css({ margin: '0px', left: disX, top: disY })
     dragDom.style.transform = "translate(0px, 0px)";
     // 获取到的值带px 正则匹配替换
-    let styL, styT;
+    let styL: any, styT: any;
 
     //注意在ie中 第一次获取到的值为组件自带50% 移动之后赋值为px
     if (sty.left.includes("%")) {
@@ -76,7 +76,7 @@ async function dragModalForNaive(opts: DragModalOptions) {
     //$(dragDom).css({ margin: '0px', left: disX, top: disY })
     dragDom.style.transform = "translate(0px, 0px)";
     // 获取到的值带px 正则匹配替换
-    let styL, styT, dragDomWidth, dragDomHeight;
+    let styL: any, styT: any;
     //注意在ie中 第一次获取到的值为组件自带50% 移动之后赋值为px
     if (sty.left.includes("%")) {
       styL = +document.body.clientWidth * (+sty.left.replace(/%/g, "") / 100);